+++ client/package.json
@@ -4,9 +4,7 @@
     "eventsource-polyfill": "0.9.6",    
     "immutability-helper": "^2.6.6",
     "react": "16.8.6",
-    "react-autobind": "^1.0.6",
     "react-dom": "16.8.6",
-    "react-sortable-hoc": "^1.10.1",    
     "whatwg-fetch": "2.0.3"
   },
   "devDependencies": {
