+++ client/src/test/favicon.png
